pyproject.toml
[project]
name = "mcp-server-deepseek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.3.0",
    "openai>=1.65.5",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-deepseek = "mcp_server_deepseek:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.9.10",
]

[tool.pyright]
include = ["src"]
exclude = [
    "**/__pycache__"
]
defineConstant = { DEBUG = true }
reportMissingImports = "error"
pythonVersion = "3.13"
reportAttributeAccessIssue = "warning"
